+++ src/main/java/main/controllers/Project/ProjectUserController.java
@@ -31,7 +31,7 @@ public class ProjectUserController extends BaseController<ProjectUserDto> {
 
     @Override
     public List<ProjectUserDto> get(ProjectUserDto template) throws AqualityException {
-        if(baseUser.isFromGlobalManagement() || baseUser.getProjectUser(template.getProject_id()).isViewer()){
+        if(baseUser.isFromGlobalManagement() || baseUser.getProjectUser(template.getProject_id()).isViewer() || template.getUser_id() != null){
             return fillProjectUsers(projectUserDao.searchAll(template));
         }else{
             throw new AqualityPermissionsException("Account is not allowed to view Project Users", baseUser);